requirements.txt
pyyaml == 6.0
psycopg2 == 2.9.5
psycopg2-binary == 2.9.5
sqlvalidator == 0.0.19
